[iOS] cocos2dx and CocosDenshion now are built for 32/64-bit.

+++ cocos2dx/include/ccTypeInfo.h
@@ -41,7 +41,7 @@ public:
 
 static inline unsigned int getHashCodeByString(const char *key)
 {
-	unsigned int len = strlen(key);
+	unsigned int len = static_cast<unsigned int>(strlen(key));
 	const char *end=key+len;
 	unsigned int hash;
